+++ src/tasks/broker/authorize-register.js
@@ -1,15 +1,42 @@
 /**
  * @param {BrokerTasks#authenticateClient} authenticateClient Task
+ * @param {RealmLimitsRepository} realmLimitsRepository Realm limits repository
+ * @param {RealtimeConnectionRepository} realtimeConnectionRepository Realtime connection repository
  * @returns {BrokerTasks#authorizeRegister} Task
  */
-module.exports = ({authenticateClient}) =>
+module.exports = ({
+  authenticateClient,
+  realmLimitsRepository,
+  realtimeConnectionRepository,
+}) =>
   /**
    * Authorize client connection
    * @typedef {Function} BrokerTasks#authorizeRegister
    * @param {string} clientId
-   * @returns {Promise<boolean>} authorization status
+   * @returns {Promise<{authorized: boolean, realmLimits?: RealmLimitsModel}>} authorization status
    */
   async clientId => {
     const client = await authenticateClient(clientId);
-    return (client || {}).authenticated === true;
+    if (!client || !client.authenticated) {
+      return {authorized: false};
+    }
+
+    const [realmLimits, numberOfRealmConnections] = await Promise.all([
+      realmLimitsRepository.findOneByRealmId(client.realmId),
+      realtimeConnectionRepository.countByRealmId(client.realmId),
+    ]);
+
+    if (
+      realmLimits
+      && Number.isFinite(realmLimits.maxConnections)
+      && realmLimits.maxConnections > 0
+      && realmLimits.maxConnections <= numberOfRealmConnections
+    ) {
+      return {authorized: false};
+    }
+
+    return {
+      authorized: true,
+      realmLimits,
+    };
   };